OPEN - issue CL-8742: [Server] Update CreatePendingPaymentJob to process o2 customers. Implementation of a new payment system http://jira.dev.now-technologies.mobi:8181/browse/CL-8742
bug with WeeklyUpdate job fixed

+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/repository/UserRepository.java
@@ -155,7 +155,7 @@ public interface UserRepository extends PagingAndSortingRepository<User, Integer
 	@Query("select u from User u " +
 			"where u.status=10 " +
 			"and u.nextSubPayment<?1")
-	List<User> getListOfUsersForWeeklyUpdate(long epochSeconds, Pageable pageable);
+	List<User> getListOfUsersForWeeklyUpdate(int epochSeconds, Pageable pageable);
 
 	@Modifying
 	@Query(value="update User user " +
